redtree/requirements.txt
amqp
billiard==3.6.3.0
boto3==1.7.52
botocore==1.10.61
cairocffi==0.8.1
CairoSVG==1.0.22
celery==4.4.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
contextlib2==0.5.5
cssselect2==0.2.1
cron_descriptor==1.2.24
croniter==0.3.32
Django==1.11.29
django-celery-beat==2.0.0
django-timezone-field
django-celery-results==1.0.1
django-crontab==0.7.1
django-markdownx==2.0.22
django-markup==1.2
django-redis==4.10.0
django-wkhtmltopdf==3.2.0
djangorestframework==3.8.1
dnspython==1.16.0
docutils==0.14
EasyProcess==0.2.3
futures==3.2.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
ipaddress==1.0.22
ipaddr==2.2.0
ipwhois==1.1.0
jmespath==0.9.3
kombu==4.6.9
lxml==4.1.1
Markdown==2.6.11
netaddr==0.7.19
pdfkit==0.6.1
pdfrw==0.4
Pillow==5.2.0
psycopg2-binary==2.7.5
pycparser==2.18
Pyphen==0.9.4
python-dateutil==2.7.3
pytz==2018.5
PyPDF2==1.26.0
PyVirtualDisplay==0.2.1
raven==6.10.0
redis==3.5.2
requests==2.19.1
s3transfer==0.1.13
selenium==3.13.0
six==1.11.0
slacker==0.9.65
slacker-log-handler==1.7.1
sentry-sdk
tinycss2==0.6.1
urllib3==1.23
vine==1.3.0
WeasyPrint==0.42.2
webencodings==0.5.1
wkhtmltopdf==0.2
MarkupSafe==1.1.1
coreapi==2.3.3
coreschema==0.0.4
itypes==1.1.0
jinja2==2.10
pygments==2.3.1
pyyaml==5.1
uritemplate==3.0.0
IPy==1.0.0
